Separate packet encoding/decoding from IO
